requirements.txt
mlflow>=2.21.2 
streamlit==1.44.1 
databricks-sdk 
pdfplumber 
load_dotenv 
databricks-cli 
databricks-vector-search-client 
pypdf 
langchain 
pandas 
requests
